# Vectorize `remove_copy` for 4 and 8 byte elements