--- a/benchmarks/src/remove.cpp
+++ b/benchmarks/src/remove.cpp
@@ -53,13 +53,9 @@ BENCHMARK(r<alg_type::rng, std::uint16_t>);
 BENCHMARK(r<alg_type::rng, std::uint32_t>);
 BENCHMARK(r<alg_type::rng, std::uint64_t>);
 
-BENCHMARK(rc<alg_type::std_fn, std::uint8_t>);
-BENCHMARK(rc<alg_type::std_fn, std::uint16_t>);
 BENCHMARK(rc<alg_type::std_fn, std::uint32_t>);
 BENCHMARK(rc<alg_type::std_fn, std::uint64_t>);
 
-BENCHMARK(rc<alg_type::rng, std::uint8_t>);
-BENCHMARK(rc<alg_type::rng, std::uint16_t>);
 BENCHMARK(rc<alg_type::rng, std::uint32_t>);
 BENCHMARK(rc<alg_type::rng, std::uint64_t>);