--- a/benchmarks/src/remove.cpp
+++ b/benchmarks/src/remove.cpp
@@ -27,7 +27,6 @@ void r(benchmark::State& state) {
     }
 }
 
-
 template <alg_type Type, class T>
 void rc(benchmark::State& state) {
     std::vector<T, not_highly_aligned_allocator<T>> src(lorem_ipsum.begin(), lorem_ipsum.end());