--- a/stl/inc/algorithm
+++ b/stl/inc/algorithm
@@ -47,8 +47,6 @@ __declspec(noalias) void __cdecl __std_reverse_copy_trivially_copyable_4(
 __declspec(noalias) void __cdecl __std_reverse_copy_trivially_copyable_8(
     const void* _First, const void* _Last, void* _Dest) noexcept;
 
-void* __stdcall __std_remove_copy_1(const void* _First, const void* _Last, void* _Dest, uint8_t _Val) noexcept;
-void* __stdcall __std_remove_copy_2(const void* _First, const void* _Last, void* _Dest, uint16_t _Val) noexcept;
 void* __stdcall __std_remove_copy_4(const void* _First, const void* _Last, void* _Dest, uint32_t _Val) noexcept;
 void* __stdcall __std_remove_copy_8(const void* _First, const void* _Last, void* _Dest, uint64_t _Val) noexcept;
 
@@ -114,10 +112,6 @@ _OutTy* _Remove_copy_vectorized(
 #else
         return reinterpret_cast<_OutTy*>(::__std_remove_copy_4(_First, _Last, _Dest, reinterpret_cast<uint32_t>(_Val)));
 #endif
-    } else if constexpr (sizeof(_InTy) == 1) {
-        return reinterpret_cast<_OutTy*>(::__std_remove_copy_1(_First, _Last, _Dest, static_cast<uint8_t>(_Val)));
-    } else if constexpr (sizeof(_InTy) == 2) {
-        return reinterpret_cast<_OutTy*>(::__std_remove_copy_2(_First, _Last, _Dest, static_cast<uint16_t>(_Val)));
     } else if constexpr (sizeof(_InTy) == 4) {
         return reinterpret_cast<_OutTy*>(::__std_remove_copy_4(_First, _Last, _Dest, static_cast<uint32_t>(_Val)));
     } else if constexpr (sizeof(_InTy) == 8) {
@@ -268,6 +262,21 @@ template <class _Iter, class _Ty1, class _Ty2>
 constexpr bool _Vector_alg_in_ranges_replace_is_safe =
     _Vector_alg_in_replace_is_safe<_Iter, _Ty1> // can search and replace
     && _Vector_alg_in_find_is_safe_elem<_Ty2, _Iter_value_t<_Iter>>; // replacement fits
+
+// Can we activate the vector algorithms for find/count?
+template <class _InIt, class _OutIt, class _Ty>
+constexpr bool _Vector_alg_in_remove_copy_is_safe =
+    // The iterators must be contiguous so we can get raw pointers.
+    _Iterators_are_contiguous<_InIt, _OutIt>
+    // The iterators must not be volatile.
+    && !_Iterator_is_volatile<_InIt>
+    && !_Iterator_is_volatile<_OutIt>
+    // Same value types
+    && is_same_v<remove_const_t<_Iter_value_t<_InIt>>, _Iter_value_t<_OutIt>>
+    // AVX2 mask compatible size
+    && sizeof(_Iter_value_t<_InIt>) >= 4
+    // The type of the value to find must be compatible with the type of the elements.
+    && _Vector_alg_in_find_is_safe_elem<_Ty, _Iter_value_t<_InIt>>;
 _STD_END
 #endif // _USE_STD_VECTOR_ALGORITHMS
 
@@ -4593,6 +4602,30 @@ _CONSTEXPR20 _OutIt remove_copy(_InIt _First, _InIt _Last, _OutIt _Dest, const _
     auto _UFirst      = _STD _Get_unwrapped(_First);
     const auto _ULast = _STD _Get_unwrapped(_Last);
     auto _UDest       = _STD _Get_unwrapped_unverified(_Dest);
+
+#if _USE_STD_VECTOR_ALGORITHMS
+    if constexpr (_Vector_alg_in_remove_copy_is_safe<decltype(_UFirst), decltype(_UDest), _Ty>) {
+        if (!_STD _Is_constant_evaluated()) {
+            if (!_STD _Could_compare_equal_to_value_type<decltype(_UFirst)>(_Val)) {
+                _STD _Seek_wrapped(_Dest, _Copy_unchecked(_UFirst, _ULast, _UDest));
+                return _Dest;
+            }
+
+            const auto _Dest_ptr = _STD _To_address(_UDest);
+            const auto _Result =
+                _STD _Remove_copy_vectorized(_STD _To_address(_UFirst), _STD _To_address(_ULast), _Dest_ptr, _Val);
+            if constexpr (is_pointer_v<decltype(_UDest)>) {
+                _UDest = _Result;
+            } else {
+                _UDest += _Result - _Dest_ptr;
+            }
+
+            _STD _Seek_wrapped(_Dest, _UDest);
+            return _Dest;
+        }
+    }
+#endif // _USE_STD_VECTOR_ALGORITHMS
+
     for (; _UFirst != _ULast; ++_UFirst) {
         if (!(*_UFirst == _Val)) {
             *_UDest = *_UFirst;