… tests/std/tests/VSO_0000000_vector_algorithms/test.cpp