@muellerj2 2024-11-06
Benchmark results on a Ryzen 7840HS notebook:

| Benchmark | main branch | this PR |
| ------------ | -------------  | -------- |  
| r<alg_type::std_fn, std::uint8_t> | 220 ns | 220 ns |
| r<alg_type::std_fn, std::uint16_t> | 250 ns | 251 ns |
| r<alg_type::std_fn, std::uint32_t> | 330 ns | 330 ns |
| r<alg_type::std_fn, std::uint64_t> | 781 ns | 806 ns |
| r<alg_type::rng, std::uint8_t> | 220 ns | 214 ns |
| r<alg_type::rng, std::uint16_t> | 261 ns | 243 ns |
| r<alg_type::rng, std::uint32_t> | 332 ns | 336 ns |
| r<alg_type::rng, std::uint64_t> | 872 ns | 816 ns |
| rc<alg_type::std_fn, std::uint32_t> | 1365 ns | 1413 ns |
| rc<alg_type::std_fn, std::uint64_t> | 942 ns | 1650 ns |
| rc<alg_type::rng, std::uint32_t> | 921 ns | 1413 ns |
| rc<alg_type::rng, std::uint64_t> | 903 ns | 1674 ns |