@AlexGuteniev 2024-11-06
> Benchmark results on a Ryzen 7840HS notebook:

Thanks. Apparently this is not the way to go 😿