@StephanTLavavej 2024-11-13
Thanks @muellerj2 - I also confirm that this is a pessimization on my desktop 5950X (Zen 3):

Benchmark                             |  Before |   After | "Speedup" (less than 1.0 is slower)
--------------------------------------|--------:|--------:|----------
`r<alg_type::std_fn, std::uint8_t>`   |  355 ns |  366 ns | 0.97
`r<alg_type::std_fn, std::uint16_t>`  |  354 ns |  335 ns | 1.06
`r<alg_type::std_fn, std::uint32_t>`  |  394 ns |  401 ns | 0.98
`r<alg_type::std_fn, std::uint64_t>`  | 1097 ns | 1122 ns | 0.98
`r<alg_type::rng, std::uint8_t>`      |  352 ns |  359 ns | 0.98
`r<alg_type::rng, std::uint16_t>`     |  335 ns |  336 ns | 1.00
`r<alg_type::rng, std::uint32_t>`     |  394 ns |  401 ns | 0.98
`r<alg_type::rng, std::uint64_t>`     | 1098 ns | 1108 ns | 0.99
`rc<alg_type::std_fn, std::uint32_t>` | 1186 ns | 1679 ns | 0.71
`rc<alg_type::std_fn, std::uint64_t>` | 1223 ns | 2056 ns | 0.59
`rc<alg_type::rng, std::uint32_t>`    | 1177 ns | 1688 ns | 0.70
`rc<alg_type::rng, std::uint64_t>`    | 1768 ns | 2077 ns | 0.85

@AlexGuteniev Do you want to rework or abandon this strategy?