@AlexGuteniev 2024-11-13
I see no rework possible.

We have the following options, besides abandoning:
 * Accept as is, in hopes future CPUs will get better, and appealing to the fact that Intel speedup is greater than AMD slowdown
 * Implement vendor detection. This takes using more state than just `__isa_available` and doing at least some detection not in VCRuntime, also will add up to the number of possible configurations.

I expect that none of these are acceptable, but leaving the final decision to you.